chore(commands): move main cmd into command folder
